pyproject.toml
[project]
name = "neonblue"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.123.5",
    "gunicorn>=23.0.0",
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "redis>=7.1.0",
    "sqlalchemy>=2.0.44",
    "uvicorn[standard]>=0.38.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
]
